Add aria-labels to doc section headings
