+++ app/assets/javascripts/src/a11y.js
@@ -1,0 +1,3 @@
+for (const anchor of document.querySelectorAll('.sect1 .anchor')) {
+  anchor.setAttribute('aria-label', 'Anchor')
+}
